+++ TDS_Server_DB/Entity/LobbyEntities/LobbyWeapons.cs
@@ -8,7 +8,7 @@ namespace TDS_Server_DB.Entity.LobbyEntities
         public WeaponHash Hash { get; set; }
         public int Lobby { get; set; }
         public int Ammo { get; set; }
-        public short? Damage { get; set; }
+        public float? Damage { get; set; }
         public float? HeadMultiplicator { get; set; }
 
         public virtual Weapons HashNavigation { get; set; }
